requirements.txt
mysql-connector-python==8.0.28
peewee==3.14.8
protobuf==3.19.4
psycopg2-binary==2.9.3
PyMySQL==1.0.2
python-dotenv==0.19.2
